Worked on adding Firebase Email/Pass Login and Signup
Login seems to be working, although I'm having issues using the same paper-inputs for the Sign Up function.  Seems like it's probably a decently simple syntax issue with how I set up the Polymer elements / functions.
Logout button added also, my plan is to have that hidden or not based on if a user is logged in.
I changed the text on the Google sign in button to "Sign in with Google Account" to make that more clear and that seems to be working fine from the work that David did. I'd like to add a Google logo/icon to that button but that's aethetics and can be done later...